Move all code into an `sdc` package
I was planning to add some more stuff at the root level so putting it
makes sense to put it into a package.
