--- scripts/loadtest-scripts/clients/__init__.py
@@ -1,1 +1,0 @@
-from clients.sdcclient import SDCClient
