+++ scripts/loadtest-scripts/sdc/clients/__init__.py
@@ -1,0 +1,1 @@
+from sdc.clients.sdcclient import SDCClient
