+++ scripts/loadtest-scripts/sdc/clients/http/authenticatedhttpclient.py
